refactor: break up monolithic dashboard page into focused components
Extract the 47KB page.tsx into a components/dashboard/ directory with
dedicated components (PipelineStatsBar, UsagePanel, ClusterSummary,
ActiveSessions, RecentTasks, PodsList, WelcomeHero, EmptyState) and a
useDashboardData hook. The page is now a 3.6KB thin layout shell that
composes extracted components. No behavior changes.
