+++ apps/web/src/components/dashboard/index.ts
@@ -1,0 +1,9 @@
+export { PipelineStatsBar } from "./pipeline-stats-bar.js";
+export { UsagePanel } from "./usage-panel.js";
+export { ClusterSummary } from "./cluster-summary.js";
+export { ActiveSessions } from "./active-sessions.js";
+export { RecentTasks } from "./recent-tasks.js";
+export { PodsList } from "./pods-list.js";
+export { WelcomeHero } from "./welcome-hero.js";
+export { EmptyState } from "./empty-state.js";
+export type { TaskStats, UsageData, MetricsHistoryPoint } from "./types.js";
